+++ HOTS4thewin/sond.h
@@ -3,12 +3,12 @@ class sond
 {
 public:
 	sond();
-	sond(int * input);
+	sond(float * input);
 	~sond();
 	float sond::get_output();
 	void sond::update();
 private :
-	int * input;
-	int output;
+	float * input;
+	float output;
 };
